Update budget amoun to BigInt

+++ models/budget.js
@@ -4,7 +4,7 @@ module.exports = (sequelize, DataTypes) => {
   var Budget = sequelize.define('Budget', {
     month: DataTypes.INTEGER,
     year: DataTypes.INTEGER,
-    amount: DataTypes.INTEGER,
+    amount: DataTypes.BIGINT,
   });
   Budget.prototype.nameMonth=function(month) {
     return MonthHelper.ConvertMonth(month)
